+++ backend/command/infrastructure/src/controllers.rs
@@ -1,27 +1,17 @@
-use std::{str::FromStr, sync::Arc};
+pub mod volunteer;
+pub mod group;
+pub mod participant;
 
-use axum::{extract::State, http::StatusCode, response::IntoResponse, routing::post, Json, Router};
-use chrono::NaiveDate;
-use lambda_http::Body;
+use std::sync::Arc;
+use axum::{routing::post, Router};
 use serde::{Deserialize, Serialize};
 use sqlx::MySqlPool;
 use tokio::sync::RwLock;
 use utoipa::ToSchema;
-
-use command_repository::user_account::{
-    group::GroupUserRepository, participant::ParticipantUserRepository,
-};
-use domain::model::{
-    gender::{gender_from_i8, Gender},
-    region::Region,
-    user_account::{
-        user_id::UserId, user_name::UserName, user_name_furigana::UserNameFurigana,
-        user_phone::UserPhone,
-    },
-};
+use lambda_http::Body;
 
 use crate::user_account::{group::GroupAccountImpl, participant::ParticipantAccountImpl};
-
+// use crate::activities::volunteer::VolunteerImpl;
 /// 失敗時のAPIレスポンスのボディを表す構造体
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct WriteApiResponseFailureBody {
@@ -34,95 +24,6 @@ pub struct WriteApiResponseSuccessBody {
     pub message: String,
 }
 
-/// グループアカウントの作成時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct CreateGroupAccountRequestBody {
-    #[schema(required = true)]
-    pub gid: String,
-    #[schema(required = true)]
-    pub name: String,
-    #[schema(required = true)]
-    pub furigana: String,
-    #[schema(required = true)]
-    pub phone: String,
-    #[schema(required = true)]
-    pub address: String,
-    #[schema(required = true)]
-    pub contents: String,
-}
-
-/// グループアカウントの更新時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct UpdateGroupAccountRequestBody {
-    #[schema(required = true)]
-    pub gid: String,
-    #[schema(required = true)]
-    pub name: String,
-    #[schema(required = true)]
-    pub furigana: String,
-    #[schema(required = true)]
-    pub phone: String,
-    #[schema(required = true)]
-    pub address: String,
-    #[schema(required = true)]
-    pub contents: String,
-}
-
-/// グループアカウントの削除時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct DeleteGroupAccountRequestBody {
-    #[schema(required = true)]
-    pub gid: String,
-}
-
-/// 参加者アカウントの作成時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct CreateParticipantAccountRequestBody {
-    #[schema(required = true)]
-    pub pid: String,
-    #[schema(required = true)]
-    pub name: String,
-    #[schema(required = true)]
-    pub furigana: String,
-    #[schema(required = true)]
-    pub phone: String,
-    #[schema(required = true)]
-    pub gender: i8,
-    #[schema(required = true, value_type = String, example = "2002-06-22")]
-    pub birthday: NaiveDate,
-    #[schema(required = true)]
-    pub region: Vec<String>,
-    #[schema(required = true)]
-    pub profile: String,
-}
-
-/// 参加者アカウントの更新時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct UpdateParticipantAccountRequestBody {
-    #[schema(required = true)]
-    pub pid: String,
-    #[schema(required = true)]
-    pub name: String,
-    #[schema(required = true)]
-    pub furigana: String,
-    #[schema(required = true)]
-    pub phone: String,
-    #[schema(required = true)]
-    pub gender: i8,
-    #[schema(required = true, value_type = String, example = "2002-06-22")]
-    pub birthday: NaiveDate,
-    #[schema(required = true)]
-    pub region: Vec<String>,
-    #[schema(required = true)]
-    pub profile: String,
-}
-
-/// 参加者アカウントの削除時のリクエストボディを表す構造体
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct DeleteParticipantAccountRequestBody {
-    #[schema(required = true)]
-    pub pid: String,
-}
 
 /// アプリケーションの状態を表す構造体
 pub struct AppState {
@@ -141,552 +42,6 @@ impl AppState {
 
 /// アプリケーションの状態を保持するための型エイリアス
 pub type AppData = Arc<RwLock<AppState>>;
-
-#[utoipa::path(
-    post,
-    path="/group-account/create",
-    request_body=CreateGroupAccountRequestBody,
-    responses(
-        (status=200, description="Create group account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Create group account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-pub async fn create_group_account(
-    State(state): State<AppData>,
-    Json(body): Json<CreateGroupAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.group_account_repository;
-
-    let gid: UserId = match UserId::from_str(&body.gid) {
-        Ok(gid) => gid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let name: UserName = match UserName::from_str(&body.name) {
-        Ok(name) => name,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let furigana: UserNameFurigana = match UserNameFurigana::from_str(&body.furigana) {
-        Ok(furigana) => furigana,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let phone: UserPhone = match UserPhone::from_str(&body.phone) {
-        Ok(phone) => phone,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let address: String = body.address;
-
-    let contents: String = body.contents;
-
-    match repository
-        .create(gid, name, furigana, phone, address, contents)
-        .await
-    {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Create group account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
-
-#[utoipa::path(
-    post,
-    path="/group-account/update",
-    request_body=UpdateGroupAccountRequestBody,
-    responses(
-        (status=200, description="Update group account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Update group account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-pub async fn update_group_account(
-    State(state): State<AppData>,
-    Json(body): Json<UpdateGroupAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.group_account_repository;
-
-    let gid: UserId = match UserId::from_str(&body.gid) {
-        Ok(gid) => gid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let name: UserName = match UserName::from_str(&body.name) {
-        Ok(name) => name,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let furigana: UserNameFurigana = match UserNameFurigana::from_str(&body.furigana) {
-        Ok(furigana) => furigana,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let phone: UserPhone = match UserPhone::from_str(&body.phone) {
-        Ok(phone) => phone,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let address: String = body.address;
-
-    let contents: String = body.contents;
-
-    match repository
-        .update(gid, name, furigana, phone, address, contents)
-        .await
-    {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Update group account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
-
-#[utoipa::path(
-    post,
-    path="/group-account/delete",
-    request_body=DeleteGroupAccountRequestBody,
-    responses(
-        (status=200, description="Delete group account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Delete group account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-pub async fn delete_group_account(
-    State(state): State<AppData>,
-    Json(body): Json<DeleteGroupAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.group_account_repository;
-
-    let gid: UserId = match UserId::from_str(&body.gid) {
-        Ok(gid) => gid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    match repository.delete(gid).await {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Delete group account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
-
-#[utoipa::path(
-    post,
-    path="/participant-account/create",
-    request_body=CreateParticipantAccountRequestBody,
-    responses(
-        (status=200, description="Create participant account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Create participant account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-async fn create_participant_account(
-    State(state): State<AppData>,
-    Json(body): Json<CreateParticipantAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.participant_account_repository;
-
-    let pid: UserId = match UserId::from_str(&body.pid) {
-        Ok(pid) => pid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let name: UserName = match UserName::from_str(&body.name) {
-        Ok(name) => name,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let furigana: UserNameFurigana = match UserNameFurigana::from_str(&body.furigana) {
-        Ok(furigana) => furigana,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let phone: UserPhone = match UserPhone::from_str(&body.phone) {
-        Ok(phone) => phone,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let gender: Gender = match gender_from_i8(&body.gender) {
-        Ok(gender) => gender,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let birthday: NaiveDate = body.birthday;
-
-    let region: Vec<Region> = body
-        .region
-        .iter()
-        .map(|r: &String| Region::from_str(r).unwrap())
-        .collect::<Vec<Region>>();
-
-    let profile: String = body.profile;
-
-    match repository
-        .create(
-            pid, name, furigana, phone, gender, birthday, region, profile,
-        )
-        .await
-    {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Create participant account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
-
-#[utoipa::path(
-    post,
-    path="/participant-account/update",
-    request_body=UpdateParticipantAccountRequestBody,
-    responses(
-        (status=200, description="Update participant account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Update participant account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-async fn update_participant_account(
-    State(state): State<AppData>,
-    Json(body): Json<UpdateParticipantAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.participant_account_repository;
-
-    let pid: UserId = match UserId::from_str(&body.pid) {
-        Ok(pid) => pid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let name: UserName = match UserName::from_str(&body.name) {
-        Ok(name) => name,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let furigana: UserNameFurigana = match UserNameFurigana::from_str(&body.furigana) {
-        Ok(furigana) => furigana,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let phone: UserPhone = match UserPhone::from_str(&body.phone) {
-        Ok(phone) => phone,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let gender: Gender = match gender_from_i8(&body.gender) {
-        Ok(gender) => gender,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    let birthday: NaiveDate = body.birthday;
-
-    let region: Vec<Region> = body
-        .region
-        .iter()
-        .map(|r: &String| Region::from_str(r).unwrap())
-        .collect::<Vec<Region>>();
-
-    let profile: String = body.profile;
-
-    match repository
-        .update(
-            pid, name, furigana, phone, gender, birthday, region, profile,
-        )
-        .await
-    {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Update participant account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
-
-#[utoipa::path(
-    post,
-    path="/participant-account/delete",
-    request_body=DeleteParticipantAccountRequestBody,
-    responses(
-        (status=200, description="Delete participant account successfully.", body=WriteApiResponseSuccessBody),
-        (status=500, description="Delete participant account failed.", body=WriteApiResponseFailureBody)
-    )
-)]
-async fn delete_participant_account(
-    State(state): State<AppData>,
-    Json(body): Json<DeleteParticipantAccountRequestBody>,
-) -> impl IntoResponse {
-    let mut lock = state.write().await;
-    let repository = &mut lock.participant_account_repository;
-
-    let pid: UserId = match UserId::from_str(&body.pid) {
-        Ok(pid) => pid,
-        Err(error) => {
-            log::warn!("error = {}", error);
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response();
-        }
-    };
-
-    match repository.delete(pid).await {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(WriteApiResponseSuccessBody {
-                message: "Delete participant account successfully.".to_string(),
-            }),
-        )
-            .into_response(),
-        Err(error) => {
-            log::error!("error = {}", error);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(WriteApiResponseFailureBody {
-                    message: error.to_string(),
-                }),
-            )
-                .into_response()
-        }
-    }
-}
 
 /// APIエンドポイントを表す列挙型
 pub enum Endpoints {
@@ -711,7 +66,7 @@ impl Endpoints {
     }
 }
 
-/// [Router] を生成する関数
+
 pub fn create_router(pool: MySqlPool) -> Router {
     // Lambdaで動かす場合
     // pub fn create_router(pool: MySqlPool) -> Router<(), Body> {
@@ -720,27 +75,27 @@ pub fn create_router(pool: MySqlPool) -> Router {
     let router = Router::new()
         .route(
             Endpoints::CreateGroupAccount.as_str(),
-            post(create_group_account),
+            post(group::create_group_account),
         )
         .route(
             Endpoints::UpdateGroupAccount.as_str(),
-            post(update_group_account),
+            post(group::update_group_account),
         )
         .route(
             Endpoints::DeleteGroupAccount.as_str(),
-            post(delete_group_account),
+            post(group::delete_group_account),
         )
         .route(
             Endpoints::CreateParticipantAccount.as_str(),
-            post(create_participant_account),
+            post(participant::create_participant_account),
         )
         .route(
             Endpoints::UpdateParticipantAccount.as_str(),
-            post(update_participant_account),
+            post(participant::update_participant_account),
         )
         .route(
             Endpoints::DeleteParticipantAccount.as_str(),
-            post(delete_participant_account),
+            post(participant::delete_participant_account),
         )
         .with_state(state);
